mvn common logical to ExternalFileTableValuedFunction
