+++ fe/fe-core/src/main/java/org/apache/doris/tablefunction/ExternalFileTableValuedFunction.java
@@ -18,12 +18,16 @@
 package org.apache.doris.tablefunction;
 
 import org.apache.doris.analysis.BrokerDesc;
+import org.apache.doris.analysis.TupleDescriptor;
 import org.apache.doris.catalog.Column;
 import org.apache.doris.catalog.Env;
 import org.apache.doris.catalog.PrimitiveType;
 import org.apache.doris.common.AnalysisException;
 import org.apache.doris.common.UserException;
 import org.apache.doris.common.util.BrokerUtil;
+import org.apache.doris.planner.PlanNodeId;
+import org.apache.doris.planner.ScanNode;
+import org.apache.doris.planner.external.ExternalFileScanNode;
 import org.apache.doris.proto.InternalService;
 import org.apache.doris.proto.InternalService.PFetchTableSchemaRequest;
 import org.apache.doris.proto.Types.PScalarType;
@@ -33,45 +37,72 @@ import org.apache.doris.rpc.BackendServiceProxy;
 import org.apache.doris.rpc.RpcException;
 import org.apache.doris.system.Backend;
 import org.apache.doris.thrift.TBrokerFileStatus;
+import org.apache.doris.thrift.TFileAttributes;
 import org.apache.doris.thrift.TFileFormatType;
+import org.apache.doris.thrift.TFileRangeDesc;
+import org.apache.doris.thrift.TFileScanRange;
+import org.apache.doris.thrift.TFileScanRangeParams;
+import org.apache.doris.thrift.TFileTextScanRangeParams;
 import org.apache.doris.thrift.TFileType;
 import org.apache.doris.thrift.TNetworkAddress;
 import org.apache.doris.thrift.TPrimitiveType;
 import org.apache.doris.thrift.TStatusCode;
 
 import com.google.common.collect.Lists;
+import com.google.protobuf.ByteString;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 import org.apache.thrift.TException;
+import org.apache.thrift.TSerializer;
 
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 
+/**
+ * ExternalFileTableValuedFunction is used for S3/HDFS/LOCAL table-valued-function
+ */
 public abstract class ExternalFileTableValuedFunction extends TableValuedFunctionIf {
     public static final Logger LOG = LogManager.getLogger(ExternalFileTableValuedFunction.class);
+    public static final String DEFAULT_COLUMN_SEPARATOR = ",";
+    public static final String DEFAULT_LINE_DELIMITER = "\n";
 
     protected List<Column> columns = null;
     protected List<TBrokerFileStatus> fileStatuses = Lists.newArrayList();
+    protected Map<String, String> locationProperties;
 
-    public abstract PFetchTableSchemaRequest getFetchTableStructureRequest() throws AnalysisException, TException;
+    protected TFileFormatType fileFormatType;
+    protected String headerType = "";
 
-    public abstract TFileFormatType getTFileFormatType();
+    protected String columnSeparator = DEFAULT_COLUMN_SEPARATOR;
+    protected String lineDelimiter = DEFAULT_LINE_DELIMITER;
 
     public abstract TFileType getTFileType();
-
-    public abstract Map<String, String> getLocationProperties();
-
-    public abstract String getColumnSeparator();
-
-    public abstract String getLineSeparator();
 
     public abstract String getFilePath();
 
     public abstract BrokerDesc getBrokerDesc();
 
-    public abstract String getHeaderType();
+    public TFileFormatType getTFileFormatType() {
+        return fileFormatType;
+    }
+
+    public Map<String, String> getLocationProperties() {
+        return locationProperties;
+    }
+
+    public String getColumnSeparator() {
+        return columnSeparator;
+    }
+
+    public String getLineSeparator() {
+        return lineDelimiter;
+    }
+
+    public String getHeaderType() {
+        return headerType;
+    }
 
     public void parseFile() throws UserException {
         String path = getFilePath();
@@ -81,6 +112,11 @@ public abstract class ExternalFileTableValuedFunction extends TableValuedFunctio
 
     public List<TBrokerFileStatus> getFileStatuses() {
         return fileStatuses;
+    }
+
+    @Override
+    public ScanNode getScanNode(PlanNodeId id, TupleDescriptor desc) {
+        return new ExternalFileScanNode(id, desc);
     }
 
     @Override
@@ -147,5 +183,51 @@ public abstract class ExternalFileTableValuedFunction extends TableValuedFunctio
                 columns.add(new Column(colName, PrimitiveType.fromThrift(tPrimitiveType), true));
             }
         }
+    }
+
+    private PFetchTableSchemaRequest getFetchTableStructureRequest() throws AnalysisException, TException {
+        // set TFileScanRangeParams
+        TFileScanRangeParams fileScanRangeParams = new TFileScanRangeParams();
+        fileScanRangeParams.setFileType(getTFileType());
+        fileScanRangeParams.setFormatType(fileFormatType);
+        fileScanRangeParams.setProperties(locationProperties);
+        fileScanRangeParams.setFileAttributes(getFileAttributes());
+
+        // get first file, used to parse table schema
+        TBrokerFileStatus firstFile = null;
+        for (TBrokerFileStatus fileStatus : fileStatuses) {
+            if (fileStatus.isIsDir()) {
+                continue;
+            }
+            firstFile = fileStatus;
+            break;
+        }
+        if (firstFile == null) {
+            throw new AnalysisException("Can not get first file, please check s3 uri.");
+        }
+
+        // set TFileRangeDesc
+        TFileRangeDesc fileRangeDesc = new TFileRangeDesc();
+        fileRangeDesc.setPath(firstFile.getPath());
+        fileRangeDesc.setStartOffset(0);
+        fileRangeDesc.setSize(firstFile.getSize());
+        // set TFileScanRange
+        TFileScanRange fileScanRange = new TFileScanRange();
+        fileScanRange.addToRanges(fileRangeDesc);
+        fileScanRange.setParams(fileScanRangeParams);
+        return InternalService.PFetchTableSchemaRequest.newBuilder()
+                .setFileScanRange(ByteString.copyFrom(new TSerializer().serialize(fileScanRange))).build();
+    }
+
+    private TFileAttributes getFileAttributes() {
+        TFileAttributes fileAttributes = new TFileAttributes();
+        if (this.fileFormatType == TFileFormatType.FORMAT_CSV_PLAIN) {
+            TFileTextScanRangeParams fileTextScanRangeParams = new TFileTextScanRangeParams();
+            fileTextScanRangeParams.setColumnSeparator(this.columnSeparator);
+            fileTextScanRangeParams.setLineDelimiter(this.lineDelimiter);
+            fileAttributes.setTextParams(fileTextScanRangeParams);
+            fileAttributes.setHeaderType(this.headerType);
+        }
+        return fileAttributes;
     }
 }
